# references/mother-harness-master/services/orchestrator/openapi.yaml
openapi: 3.0.3
info:
  title: Mother-Harness Orchestrator API
  description: |
    The Mother-Harness Orchestrator API provides endpoints for task management,
    approval workflows, library management, and system monitoring.

    ## Authentication
    All endpoints (except /health) require JWT authentication via Bearer token.

    ## Roles
    - **user**: Can create and view their own tasks
    - **approver**: Can approve/reject approval requests
    - **admin**: Full system access
  version: 0.1.0
  contact:
    name: Mother-Harness Team
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://orchestrator.production.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /api/tasks/{taskId}:
    get:
      summary: Get task details
      description: Retrieve detailed information about a specific task
      operationId: getTask
      tags:
        - Tasks
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
          description: Task ID
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/projects:
    post:
      summary: Create a project
      description: Create a new project to group related tasks
      operationId: createProject
      tags:
        - Projects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: "E-commerce Platform"
                description:
                  type: string
                  example: "Online shopping platform development"
      responses:
        '200':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'

    get:
      summary: List projects
      description: List all projects for the authenticated user
      operationId: listProjects
      tags:
        - Projects
      parameters:
        - name: user_id
          in: query
          description: Filter by user ID (admin only)
          schema:
            type: string
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
...
  /api/approvals/{id}/respond:
    post:
      summary: Respond to approval request
      description: Approve or reject an approval request
      operationId: respondToApproval
      tags:
        - Approvals
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Approval ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - approved
              properties:
                approved:
                  type: boolean
                  description: Whether to approve or reject
                  example: true
                notes:
                  type: string
                  description: Optional notes explaining the decision
                  example: "Approved after security review"
      responses:
        '200':
          description: Response recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/libraries:
    get:
      summary: List libraries
      description: List all document libraries
      operationId: listLibraries
      tags:
        - Libraries
      parameters:
        - name: search
          in: query
          description: Search term to filter libraries
          schema:
            type: string
      responses:
        '200':
          description: List of libraries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Library'

    post:
      summary: Create library
      description: Create a new document library
      operationId: createLibrary
      tags:
        - Libraries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - folder_path
              properties:
                name:
                  type: string
                  example: "Technical Documentation"
                folder_path:
                  type: string
                  example: "/libraries/tech-docs"
                description:
                  type: string
                  example: "Internal technical documentation repository"
                auto_scan:
                  type: boolean
                  default: true
      responses:
        '200':
          description: Library created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Library'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Task:
      type: object
      properties:
        id:
          type: string
          example: "task-abc123"
        project_id:
          type: string
          example: "project-xyz"
        user_id:
          type: string
          example: "user-789"
        description:
          type: string
          example: "Research and implement user authentication"
        status:
          type: string
          enum: [planning, executing, approval_needed, completed, failed]
        todo_list:
          type: array
          items:
            $ref: '#/components/schemas/TodoItem'
        steps_completed:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
    TodoItem:
      type: object
      properties:
        id:
          type: string
          example: "step-1"
        description:
          type: string
          example: "Research authentication best practices"
        agent:
          type: string
          enum: [orchestrator, researcher, coder, design, analyst, critic, skeptic, rag, librarian, vision, update, toolsmith]
        status:
          type: string
          enum: [pending, in_progress, completed, failed]
        depends_on:
          type: array
          items:
            type: string
        require_approval:
          type: boolean
        approval_type:
          type: string
          enum: [file_write, code_execution, git_push, workflow_creation, api_call]
        risk:
          type: string
          enum: [low, medium, high]
        result:
          type: object
          description: Step execution result
        error:
          type: string
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
    Project:
      type: object
      properties:
        id:
          type: string
          example: "project-abc123"
        user_id:
          type: string
        name:
          type: string
          example: "E-commerce Platform"
        description:
          type: string
        created_at:
          type: string
          format: date-time
...
    Library:
      type: object
      properties:
        id:
          type: string
          example: "library-abc123"
        name:
          type: string
          example: "Technical Documentation"
        folder_path:
          type: string
          example: "/libraries/tech-docs"
        description:
          type: string
        auto_scan:
          type: boolean
        last_scan:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
  responses:
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Insufficient permissions"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Resource not found"
...